+++ apps/meteor/client/components/message/IgnoredContent.tsx
@@ -17,7 +17,7 @@ const IgnoredContent = ({ onShowMessageIgnored }: IgnoredContentProps): ReactEle
 	};
 
 	return (
-		<MessageBody data-qa-type='message-body'>
+		<MessageBody data-qa-type='message-body' dir='auto'>
 			<Box display='flex' alignItems='center' fontSize='c2' color='hint'>
 				<p role='button' onClick={showMessageIgnored} style={{ cursor: 'pointer' }}>
 					<Icon name='chevron-left' /> {t('Message_Ignored')}